changed id types to Long and added a Image support

+++ city_infromation_service/src/main/java/viachaslausviatski/city_infromation_service/repositories/EventRequestRepository.java
@@ -3,5 +3,5 @@ package viachaslausviatski.city_infromation_service.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import viachaslausviatski.city_infromation_service.entities.EventRequest;
 
-public interface EventRequestRepository extends JpaRepository<EventRequest,Integer> {
+public interface EventRequestRepository extends JpaRepository<EventRequest,Long> {
 }
