+++ city_infromation_service/src/main/java/viachaslausviatski/city_infromation_service/repositories/ObjectTypeRepository.java
@@ -3,5 +3,5 @@ package viachaslausviatski.city_infromation_service.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import viachaslausviatski.city_infromation_service.entities.ObjectType;
 
-public interface ObjectTypeRepository extends JpaRepository<ObjectType,Integer> {
+public interface ObjectTypeRepository extends JpaRepository<ObjectType,Long> {
 }
