+++ city_infromation_service/src/main/java/viachaslausviatski/city_infromation_service/repositories/OwnerRepository.java
@@ -3,5 +3,5 @@ package viachaslausviatski.city_infromation_service.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import viachaslausviatski.city_infromation_service.entities.Owner;
 
-public interface OwnerRepository extends JpaRepository<Owner,Integer> {
+public interface OwnerRepository extends JpaRepository<Owner,Long> {
 }
